Directional Shadows fixed

+++ DirectionalLight.h
@@ -7,7 +7,7 @@ class DirectionalLight :
 public:
 	DirectionalLight();
 	DirectionalLight(
-		GLfloat shadowWidth, GLfloat shadowHeight,
+		GLuint shadowWidth, GLuint shadowHeight,
 		GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat aIntensity, GLfloat dIntensity,
 		GLfloat xDir, GLfloat yDir, GLfloat zDir);
